Minor update to examples startup code.
Updated the startup code to use startServer instead of
runServer. Starting the server then entering in an
endless waiting loop is more compatible with running
the example as a systemd service than if we use runServer
which waits for a user input then quits. When running as
a systemd service this waiting for user input never
happens and the program quits immediately.