Made Widget operations abstract.
Refactored the Widget class to make its properties and
events abstract so they can be implemented by both basic
widgets as well as components. This eliminates the
disconnect in behavior between components and basic
widgets. So now using those basic events and properties
on components will automatically be applied to the root
widget of that component.